#13 lesson progress view - review remarks: load progress instead of create progress, LessonProgressDto creates out of LessonProgress, make it compile on Java 1.8, delete empty spaces, move types to shared
