+++ adapters/src/main/kotlin/com/krzykrucz/elesson/currentlesson/lessonprogress/adapters/persistence/LessonProgressDatabaseConnector.kt
@@ -2,12 +2,12 @@ package com.krzykrucz.elesson.currentlesson.lessonprogress.adapters.persistence
 
 import arrow.core.toOption
 import arrow.fx.IO
-import com.krzykrucz.elesson.currentlesson.lessonprogress.usecase.CreateLessonProgressView
 import com.krzykrucz.elesson.currentlesson.lessonprogress.usecase.LessonProgress
 import com.krzykrucz.elesson.currentlesson.lessonprogress.usecase.LessonProgressError
+import com.krzykrucz.elesson.currentlesson.lessonprogress.usecase.LoadLessonProgress
 import com.krzykrucz.elesson.currentlesson.monolith.Database
 
-fun createLessonProgressView(): CreateLessonProgressView = { lessonIdentifier ->
+fun loadLessonProgress(): LoadLessonProgress = { lessonIdentifier ->
     Database.LESSON_DATABASE[lessonIdentifier].toOption()
         .map { currentLesson ->
             LessonProgress(
