+++ adapters/src/main/kotlin/com/krzykrucz/elesson/currentlesson/lessonprogress/adapters/rest/LessonProgressRouteConfig.kt
@@ -16,6 +16,4 @@ class LessonProgressRouteConfig {
             }
         }
     }
-
-
 }
