+++ adapters/src/main/kotlin/com/krzykrucz/elesson/currentlesson/lessonprogress/usecase/LessonProgressPersistence.kt
@@ -8,4 +8,4 @@ sealed class LessonProgressError {
     data class LessonNotFound(val error: String = "Could not find lesson") : LessonProgressError()
 }
 
-typealias CreateLessonProgressView = (LessonIdentifier) -> IO<Either<LessonProgressError, LessonProgress>>
+typealias LoadLessonProgress = (LessonIdentifier) -> IO<Either<LessonProgressError, LessonProgress>>
