+++ adapters/src/main/kotlin/com/krzykrucz/elesson/currentlesson/monolith/Database.kt
@@ -3,8 +3,6 @@ package com.krzykrucz.elesson.currentlesson.monolith
 import arrow.core.Option
 import com.krzykrucz.elesson.currentlesson.attendance.domain.Attendance
 import com.krzykrucz.elesson.currentlesson.attendance.domain.IncompleteAttendanceList
-import com.krzykrucz.elesson.currentlesson.lessonprogress.usecase.LessonStatus
-import com.krzykrucz.elesson.currentlesson.lessonprogress.usecase.Scheduled
 import com.krzykrucz.elesson.currentlesson.shared.*
 import java.time.LocalDate
 import java.util.concurrent.ConcurrentHashMap
