+++ adapters/src/main/kotlin/com/krzykrucz/elesson/currentlesson/startlesson/adapters/persistence/StartedLessonPersistenceAdapter.kt
@@ -2,9 +2,9 @@ package com.krzykrucz.elesson.currentlesson.startlesson.adapters.persistence
 
 import arrow.core.Option
 import arrow.effects.IO
-import com.krzykrucz.elesson.currentlesson.lessonprogress.usecase.InProgress
 import com.krzykrucz.elesson.currentlesson.monolith.Database.Companion.LESSON_DATABASE
 import com.krzykrucz.elesson.currentlesson.monolith.PersistentCurrentLesson
+import com.krzykrucz.elesson.currentlesson.shared.InProgress
 import com.krzykrucz.elesson.currentlesson.shared.LessonIdentifier
 import com.krzykrucz.elesson.currentlesson.startlesson.domain.PersistStartedLessonIfDoesNotExist
 import com.krzykrucz.elesson.currentlesson.startlesson.domain.StartedLesson
